+++ src/epics/alert.ts
@@ -5,6 +5,9 @@ import { ADD_SUCCESS_ALERT, ADD_DANGER_ALERT, ADD_ALERT } from '../actionConsts'
 import { of } from 'rxjs'
 import * as R from 'ramda'
 import { Epic } from './epic'
+import type { EpicPayload } from '../types'
+
+export type ExpiresOnPayload = EpicPayload & { expiresOn: number }
 
 export class AlertEpic extends Epic {
   [ADD_ALERT](action$: any) {
@@ -12,11 +15,11 @@ export class AlertEpic extends Epic {
       ofType(ADD_DANGER_ALERT, ADD_SUCCESS_ALERT, ADD_ALERT),
       map(R.prop('payload')),
       filter(
-        (payload: EpicPayload) => R.prop('expiresOn', payload) !== undefined
+        (payload: EpicPayload): payload is ExpiresOnPayload =>
+          payload.expiresOn !== undefined
       ),
-      mergeMap((payload: EpicPayload) => {
-        // @ts-ignore
-        const timeOfAlertDismiss = R.prop('expiresOn', payload) - Date.now()
+      mergeMap((payload: EpicPayload & { expiresOn: number }) => {
+        const timeOfAlertDismiss = payload.expiresOn - Date.now()
         return of(Actions.dismissAlert(payload)).pipe(delay(timeOfAlertDismiss))
       })
     )