+++ src/epics/search.ts
@@ -5,6 +5,7 @@ import { TRIGGER_SEARCH, FETCH_SEARCH_ENTRIES } from '../actionConsts'
 import * as Logger from '../utils/Logger'
 import * as R from 'ramda'
 import { Epic } from './epic'
+import { EpicPayload } from '../types'
 
 export class SearchEpic extends Epic {
   [TRIGGER_SEARCH](action$: any) {
@@ -22,7 +23,7 @@ export class SearchEpic extends Epic {
       ofType(FETCH_SEARCH_ENTRIES),
       map(R.prop('payload')),
       map((payload: EpicPayload) => {
-        const query = this.queryBuilder.buildQuery({ queryType: 'search' })
+        const query = this.queryTool.buildQuery({ queryType: 'search' })
         const variables = {
           queryString: payload.queryString
             ? payload.queryString.replace(/[%_]/g, '\\$&')
@@ -32,7 +33,7 @@ export class SearchEpic extends Epic {
         return { queryString: payload.queryString, query, variables }
       }),
       mergeMap((context: any) =>
-        this.queryBuilder
+        this.queryTool
           .sendRequest({ query: context.query, variables: context.variables })
           .then(({ data, error }) => ({ context, data, error }))
       ),