+++ CHANGES.md
@@ -1,8 +1,21 @@
-#Conveyor-redux Changelog
+#conveyor-redux Changelog
+## Version 0.3.0
 
-## Version 0.1.1
+-- UNRELEASED --
 
-Unreleased
+Update to work with conveyor-schema typescript changes
+
+## Version 0.2.1
+
+Released 4-21-21
+
+Bugfix for DETAIL_TABLE_REMOVE_SUBMIT Epic
+
+## Version 0.2.0
+
+Released 4-15-21
+
+Update for Tooltips to follow the fieldOrder metadata
 
 ## Version 0.1.0